## Why is this change being made?
Need to add exception handler to create crash dump for fault condition.

## What is changed?
Add exception handler to NVIC with fault cases.
Add Debug Monitor flag to enable bug check
Add exception handler init
Add ThreadX stack error callback, which makes bug check
Modify Crash dump CLI to simulate ThreadX stack error
Add unit tests

## How is the change being tested?

- [ ] No test, document update only
- [x] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

Add exception handler and its init (register) with unit tests

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds an exception handler for various processor faults in the Cortex-M7, along with unit tests.
- `src/libs/exception_handler/src/exception_handler.c`: Implemented exception handler functions for different fault types and registered the main exception handler.
- `tests/unit/libs/exception_handler/test_exception_handler.cpp`: Added unit tests for the exception handler.
- `tests/unit/libs/runtime_init/exception_handler_init/test_exception_handler_init.cpp`: Added tests for the initialization of the exception handler.
- `src/libs/exception_handler/CMakeLists.txt`: Updated CMake configuration to include the new exception handler library.

Related work items: #1484988

+++ src/libs/crash_dump/inc/bug_check.h
@@ -32,8 +32,8 @@
  *      User parameters included with the bug check information.
  */
 #define BUG_CHECK(errorCode, p1, p2) \
-    DBGPRINT("Bug Check: [%"PRId32"] File [%s:%d] P1: [%"PRIu32"] P2: [%"PRIu32"]\n", (int32_t)errorCode, __FILE__, __LINE__, (uint32_t)p1, (uint32_t)p2); \
-    crash_dump_bug_check(errorCode, (uint32_t)__FILE__, __LINE__, p1, p2)
+    DBGPRINT("Bug Check: [%"PRId32"] File [%s:%d] P1: [%"PRIu32"] P2: [%"PRIu32"]\n", (int32_t)(errorCode), __FILE__, __LINE__, (uint32_t)(p1), (uint32_t)(p2)); \
+    crash_dump_bug_check((errorCode), (uint32_t)__FILE__, __LINE__, (uint32_t)(p1), (uint32_t)(p2))
 
 /**
  * The BUG_ASSERT() macro will generate a bug check if the condition parameter
